+++ web/src/classic/core/mantle/data/index.ts
@@ -6,7 +6,7 @@ import { fetchGeoRSS } from "./georss";
 import { fetchGMLData } from "./gml";
 import { fetchGPXfile } from "./gpx";
 import { fetchGTFS } from "./gtfs";
-import { fetchShapefile } from "./shapefile/shapefile";
+import { fetchShapefile } from "./shapefile";
 import { FetchOptions } from "./utils";
 
 export type DataFetcher = (
