+++ web/src/classic/core/mantle/data/shapefile/parseZip.ts
@@ -3,7 +3,8 @@ import proj4 from "proj4";
 
 import { parseDbf } from "./parseDbf";
 import { parseShp } from "./parseShp";
-import { combine } from "./shapefile";
+
+import { combine } from ".";
 
 export async function parseZip(buffer: ArrayBuffer): Promise<GeoJSON.GeoJSON | GeoJSON.GeoJSON[]> {
   const zip = await JSZip.loadAsync(buffer);
